requirements.txt
djangorestframework==3.11.0
django==2.2.2
django-cors-headers==3.2.1
Pillow==7.1.2
django-rest-swagger==2.2.0
django-debug-toolbar
python-decouple